--- example/src/CostCalculatorFlowWithExportPdf.tsx
@@ -1,101 +1,0 @@
-import {
-  CostCalculatorFlow,
-  CostCalculatorForm,
-  CostCalculatorSubmitButton,
-  CostCalculatorResetButton,
-  RemoteFlows,
-  useCostCalculatorEstimationPdf,
-  buildCostCalculatorEstimationPayload,
-} from '@remoteoss/remote-flows';
-import type {
-  CostCalculatorEstimateResponse,
-  CostCalculatorEstimationFormValues,
-} from '@remoteoss/remote-flows';
-import './App.css';
-import { useState } from 'react';
-
-const estimationOptions = {
-  title: 'Estimate for a new company',
-  includeBenefits: true,
-  includeCostBreakdowns: true,
-};
-
-function CostCalculatorFormDemo() {
-  const [estimations, setEstimations] =
-    useState<CostCalculatorEstimateResponse | null>(null);
-  const [payload, setPayload] =
-    useState<CostCalculatorEstimationFormValues | null>(null);
-
-  const exportPdfMutation = useCostCalculatorEstimationPdf();
-
-  const handleExportPdf = () => {
-    if (payload) {
-      exportPdfMutation.mutate(buildCostCalculatorEstimationPayload(payload), {
-        onSuccess: (response) => {
-          if (response?.data?.data?.content !== undefined) {
-            const a = document.createElement('a');
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            a.href = response.data.data.content as any;
-            a.download = 'estimation.pdf';
-            document.body.appendChild(a);
-            a.click();
-            document.body.removeChild(a);
-          }
-        },
-        onError: (error) => {
-          console.error({ error });
-        },
-      });
-    }
-  };
-  return (
-    <>
-      <CostCalculatorFlow
-        estimationOptions={estimationOptions}
-        render={(props) => {
-          if (props.isLoading) {
-            return <div>Loading...</div>;
-          }
-
-          return (
-            <div>
-              <CostCalculatorForm
-                onSubmit={(payload) => setPayload(payload)}
-                onError={(error) => console.error({ error })}
-                onSuccess={(response) => {
-                  setEstimations(response);
-                }}
-              />
-              <CostCalculatorSubmitButton>
-                Get estimate
-              </CostCalculatorSubmitButton>
-              <CostCalculatorResetButton>Reset</CostCalculatorResetButton>
-            </div>
-          );
-        }}
-      />
-      {estimations && <button onClick={handleExportPdf}>Export as PDF</button>}
-    </>
-  );
-}
-
-export function CostCalculatorFlowWithExportPdf() {
-  const fetchToken = () => {
-    return fetch('/api/token')
-      .then((res) => res.json())
-      .then((data) => ({
-        accessToken: data.access_token,
-        expiresIn: data.expires_in,
-      }))
-      .catch((error) => {
-        console.error({ error });
-        throw error;
-      });
-  };
-
-  return (
-    <RemoteFlows auth={() => fetchToken()}>
-      <CostCalculatorFormDemo />
-    </RemoteFlows>
-  );
-}
